% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodo.R
\name{dodo}
\alias{dodo}
\title{My First Function}
\usage{
dodo(n_sum, n_count)
}
\arguments{
\item{num_sum}{explain parameter 1}

\item{num_count}{explain parameter 2}
}
\value{
the average
}
\description{
dodo does the average calculation.
}

