% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fufu.R
\name{fufu}
\alias{fufu}
\title{weight_and_iris_graph}
\usage{
fufu(person_name, weight_kg)
}
\arguments{
\item{person_name}{顯示的人名}

\item{weight_kg}{該員的體重公斤}
}
\value{
Nothing
}
\description{
顯示體重公斤 與 iris 的資料分佈。
}

